+++ pepclibs/msr/PlatformInfo.py
@@ -39,6 +39,7 @@ _EFREQ_CPUS = CPUInfo.EMRS +         \
               CPUInfo.SANDYBRIDGES + \
               CPUInfo.WESTMERES +    \
               CPUInfo.NEHALEMS +     \
+              CPUInfo.CRESMONTS +    \
               CPUInfo.TREMONTS +     \
               CPUInfo.GOLDMONTS +    \
               CPUInfo.PHIS
@@ -59,6 +60,7 @@ _MIN_OPER_RATIO_CPUS = CPUInfo.METEORLAKES +               \
                        CPUInfo.BROADWELLS +                \
                        CPUInfo.HASWELLS +                  \
                        (CPUInfo.INTEL_FAM6_IVYBRIDGE, ) +  \
+                       CPUInfo.CRESMONTS +                 \
                        CPUInfo.TREMONTS +                  \
                        CPUInfo.GOLDMONTS +                 \
                        CPUInfo.PHIS